split lookup code into separate module
